extract out collections
